CAMEL-23703: tooling/camel-exe - split native bootstrap from camel-launcher
Move camel.exe build and tests into tooling/camel-exe so Windows CI can
validate the native bootstrap without building the full jbang plugin tree.
camel-launcher now copies the exe artifact on Windows for assembly.
