+++ tooling/camel-exe/src/main/native/camel.c
