+++ tooling/camel-exe/src/test/java/org/apache/camel/tooling/exe/CamelExeBootstrapTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.dsl.jbang.launcher;
+package org.apache.camel.tooling.exe;
 
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
