+++ tooling/pom.xml
@@ -42,6 +42,7 @@
 
     <modules>
         <module>parent</module>
+        <module>camel-exe</module>
         <module>spi-annotations</module>
         <module>camel-tooling-model</module>
         <module>camel-util-json</module>
